# cmd/nube/openapi.yaml
openapi: 3.0.3
info:
  title: Nube Server API
  version: 0.2.0
  description: >
    Multi-tenant central server for NubeIO — workspace/user management,
    app catalog, install flow, per-user MCP tool scoping, agent execution,
    app store with reviews and sharing.
servers:
  - url: http://localhost:8090
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /bootstrap:
    post:
      operationId: bootstrap
      summary: Create first workspace and admin user
      security: []
      x-cli:
        command: bootstrap
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workspaceName, adminName, adminEmail]
              properties:
                workspaceName: { type: string, description: Workspace name }
                adminName:     { type: string, description: Admin user name }
                adminEmail:    { type: string, description: Admin email address }
      responses:
        "201":
          description: Created

  # ---------------------------------------------------------------------------
  # User management
  # ---------------------------------------------------------------------------
...
  /users/{id}:
    get:
      operationId: getUser
      summary: Get user by ID (admin)
      x-cli:
        command: users get
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: User ID
      responses:
        "200":
          description: OK

    delete:
      operationId: deleteUser
      summary: Delete user (admin)
      x-cli:
        command: users delete
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: User ID
      responses:
        "200":
          description: Deleted
...
  /api/agents/sessions/{id}:
    get:
      operationId: getSession
      summary: Get session detail with full result text
      x-cli:
        command: sessions get
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Session ID
      responses:
        "200":
          description: OK

  # ---------------------------------------------------------------------------
  # Provider test
  # ---------------------------------------------------------------------------
  /api/agents/providers/{name}/test:
    post:
      operationId: testProvider
      summary: Test connectivity to a specific AI provider
      x-cli:
        command: agents test
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
          description: "Provider name (claude, ollama, openai, etc.)"
      responses:
        "200":
          description: Test result with availability, latency, and models

  # ---------------------------------------------------------------------------
  # User preferences
  # ---------------------------------------------------------------------------
  /users/me/preferences:
    get:
      operationId: getUserPreferences
      summary: Get current user's AI preferences (default provider/model)
      x-cli:
        command: preferences get
      responses:
        "200":
          description: OK
    put:
      operationId: updateUserPreferences
      summary: Set current user's default AI provider and model
      x-cli:
        command: preferences set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                default_provider: { type: string, description: "Default provider (claude, ollama, openai, etc.)" }
                default_model:    { type: string, description: "Default model (gemma3, gpt-4.1, etc.)" }
      responses:
        "200":
          description: Preferences updated

  # ---------------------------------------------------------------------------
  # Admin
  # ---------------------------------------------------------------------------
  /api/settings/providers:
    get:
      operationId: getProviderConfig
      summary: Get global provider configuration (admin)
      x-cli:
        command: admin providers
      responses:
        "200":
          description: Provider config with availability
    put:
      operationId: updateProviderConfig
      summary: Update global provider configuration (admin)
      # x-cli: admin-only, use curl or frontend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                providers:
                  type: object
                  additionalProperties:
                    type: object
                    properties:
                      enabled: { type: boolean }
                      api_key: { type: string }
                      host:    { type: string }
      responses:
        "200":
          description: Config updated
...